making plugin work asynchronously
The approach of blocking execution of a long task contradicts with
'netdata' way of work: i.e. it is designed (*sigh*) to launch single
thread for each charts.d orchestrator which runs each collector one by
one. This leads to blocking of whole charts.d instance and lack of data
as a result.